Bug 18019: Empty prompt shown after non-en-US update
Avoid displaying the "request English language web pages?" prompt
on non-HTTP pages. Previously, the prompt was displayed when the
about:tbupdate page was opened and, on Linux at least, the prompt
dialog came up blank. To avoid a similar problem in the case where
someone sets their homepage to an HTTP page, we delay showing the
prompt for two seconds when the URL of the page that is being loaded
matches one of the configured home page URLs.

Remove references to the following obsolete Firefox preferences:
  intl.accept_charsets
  intl.charsetmenu.browser.cache
Remove the following unused default preferences:
  extensions.torbutton.saved.accept_charsets
  extensions.torbutton.spoof_charset
  extensions.torbutton.spoof_language
  extensions.torbutton.spoof_locale
Remove the following unused entity from all dtd files:
  torbutton.prefs.spoof_english